requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
psycopg2-binary
python-dotenv
python-jose
passlib[bcrypt]
jinja2
pdfkit
python-multipart
